씨앗에 click 이벤트 적용+ className에맞게 ToDo 정보 띄우기
