+++ public/CSS/env_step2.css
@@ -9,3 +9,9 @@
     height: 50px;
     border: 10px solid black;
 }
+
+.seed_img{
+    display: inline-block;
+    width: 150px;
+    height: 150px;
+}
