+++ public/CSS/style2.css
@@ -1,5 +1,5 @@
 @import "reset.css";
-@import "env_step.css";
+@import "env_step2.css";
 @import "todolist.css";
 
 body {
@@ -53,6 +53,5 @@ a {
     top: 150px;
     width: 80%;
     height: 80%;
-    background-color: gray;
 }
